Refactor image source paths to use addBasePath for GitHub Pages compatibility
Images are not affected by basePath and assetPrefix. Therefore, I have
wrapped their paths using the addBasePath utility to ensure correct
resolution when deploying to GitHub Pages.

https://nextjs.org/docs/pages/api-reference/config/next-config-js/basePath

Note:
The `addBasePath` function can not be defined in `next.config.ts`, which
is only evaluated at build time by Node.js and is not part of the
Webpack bundle. As a result, importing it from in-app code (like
`/app/page.tsx`) caused a runtime error during static export.